src/mcp/requirements.txt
# Core framework
fastapi>=0.100,<0.136
sentry-sdk[fastapi]>=2.35.0
uvicorn[standard]>=0.46.0,<1
pydantic>=2.0,<3

# MCP protocol
mcp>=1.27.0,<2

# HTTP client
httpx>=0.24,<0.29

# Templating (Bifrost config generation)
jinja2>=3.1,<4

# Databases
chromadb>=0.5,<0.6  # server pinned to 0.5.23 — do not widen until server upgrade
neo4j>=5.0,<7       # neo4j 6.x OK — audit verified: all sessions use `with`, no deprecated tx APIs
redis>=4.0,<8

# AI / NLP
tiktoken>=0.5,<1
langgraph>=0.3.0,<2.0  # langgraph 1.x OK — audit verified: only StateGraph/END from langgraph.graph (triage.py)

# File parsing
pdfplumber>=0.11.9,<0.12
python-docx>=0.8,<2
openpyxl>=3.1,<4
pandas>=2.0,<3.1

# Search
bm25s>=0.3,<1.0
PyStemmer>=2.2

# Scheduling
apscheduler>=3.10,<4

# Queue (Workstream E Phase 5a — opt-in via INGEST_QUEUE_MODE=async; reuses Redis already in stack)
rq>=2.0,<3

# Markdown header-hierarchy chunker (Workstream E Phase 2b.2)
langchain-text-splitters>=0.3,<1

# Code AST chunker (Workstream E Phase 2b.3) — Python grammar today;
# tree-sitter-{javascript,typescript,go,rust} added per future sub-phase.
tree-sitter>=0.23,<1
tree-sitter-python>=0.23,<1

# Email parser (Workstream E Phase 2b.4) — RFC822 header preservation +
# reply-chain stripping. Both libs zero-transitive-dep, pure Python.
mail-parser>=4.0,<5
quotequail>=0.4,<1

# Logging (used by trading agent + internal tests)

# Auth (optional — only needed for multi-user mode)
# optional: PyJWT>=2.8,<3
# optional: bcrypt>=4.0,<6

# Utilities
cryptography>=42,<47
python-multipart>=0.0.9,<1

# Plugins (Pro tier — system deps: tesseract-ocr, ffmpeg in Dockerfile)
# optional (Pro OCR plugin): pytesseract>=0.3.10
# optional (Pro OCR plugin): Pillow>=10.0
# Vision: uses httpx (already in core) + Bifrost/OpenRouter — no extra deps


# Security — content-type sniffing for upload validation (magic-byte guard)
filetype>=1.2,<2
